items car list

+++ src/componets/actions/actionsProductos.jsx
@@ -1,15 +1,28 @@
 
-export function addProducto(id, image, nombre, cantidad) {
+export function addProducto(id, nombre, image, cantidad) {
     const array=[]
-    array.push({id, nombre, image, cantidad})
+    const datosExistentes = localStorage.getItem('listaProductos')
 
-    localStorage.setItem('listaProductos', JSON.stringify(array))
+    if(datosExistentes === null){
+        array.push({id, nombre, image, cantidad})
+        localStorage.setItem('listaProductos', JSON.stringify(array))
+    }else{
+        const auxarray =  datosExistentes
+        const parse = JSON.parse(auxarray)
+        //console.log(parse)
+        array.push(...parse, {id, nombre, image, cantidad})
+        //console.log('otro array')
+        //console.log(array)
+        localStorage.clear()
+        localStorage.setItem('listaProductos', JSON.stringify(array))
+    }
 }
 
 export function consProductos(){
-    localStorage.getItem('listaProductos')
+    const array = localStorage.getItem('listaProductos')
+    return JSON.parse(array)
 }
 
-const consCantidadArray = () => {
-
-}
+export function deleteProducto(){
+    
+}
